Testing.

+++ python/controller/wsprauto.py
@@ -850,18 +850,19 @@ class Automate:
             except Exception as e:
                 return DISP_NONRECOVERABLE_ERROR, 'Exception starting WsprryPi [%s]' % (str(e)) % (params)
         elif subcommand == WSPRRY_WAIT:
-            if self.__wsprrypi_proc.poll() == None:
-                while True:
-                    print('Waiting for WsprryPi to finish')
-                    try:
-                        # Give it 4.0 minutes to close as cycles are 2 mins and we could wait 2 mins for the start
-                        self.__wsprrypi_proc.wait(245)
-                        print('WsprryPi exited')
-                        break
-                    except subprocess.TimeoutExpired:
-                        self.__wsprrypi_proc.kill()
-                        return DISP_RECOVERABLE_ERROR, 'Timeout waiting for WsprryPi to terminate ... killing!'
-                        break
+            if self.__wsprrypi_proc != None:
+                if self.__wsprrypi_proc.poll() == None:
+                    while True:
+                        print('Waiting for WsprryPi to finish')
+                        try:
+                            # Give it 4.0 minutes to close as cycles are 2 mins and we could wait 2 mins for the start
+                            self.__wsprrypi_proc.wait(245)
+                            print('WsprryPi exited')
+                            break
+                        except subprocess.TimeoutExpired:
+                            self.__wsprrypi_proc.kill()
+                            return DISP_RECOVERABLE_ERROR, 'Timeout waiting for WsprryPi to terminate ... killing!'
+                            break
         elif subcommand == WSPRRY_KILL:
             if self.__wsprrypi_proc.poll() == None:
                 self.__wsprrypi_proc.kill()
